updated test for hotel
