+++ hotel/tests/__init__.py
@@ -1,3 +1,3 @@
 # -*- encoding: utf-8 -*-
 
-from . import test_folio
+from . import test_hotel
